requirements.txt
streamlit>=1.32.0
torch>=2.2.0
transformers>=4.37.0
Pillow>=10.2.0
numpy>=1.26.0
tqdm>=4.66.0 